server/requirements.txt
aioconsole==0.1.14
aiocontextvars==0.2.2
aiofiles==0.4.0
aioredis==1.2.0
alembic==1.0.11
async-timeout==3.0.1
asyncpg==0.18.3
certifi==2019.6.16
chardet==3.0.4
contextvars==2.4
environs==5.2.1
gino==0.8.3
h11==0.8.1
h2==3.1.1
hiredis==1.0.0
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
immutables==0.9
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.0
multidict==4.5.2
psycopg2==2.8.3
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.6.2
six==1.12.0
SQLAlchemy==1.3.7
ujson==1.35
urllib3==1.25.3
uvloop==0.13.0
websockets==6.0
